requirements.txt
functions-framework==3.*
google-cloud-storage~=2.14.0
httpx~=0.26.0